Updates to denscubeprop, wavefunction.
Denscubeprop can now generate/compute eud for alpha beta U and S type
function numbering. The wavefunction base has been modified to allow
post-scf modules to edit the Da, Db matrices. The ccdensity/dipole code
has been modified to set this density matrix when it has been computed.
This might need to be moved so that the Da,Db matrices will be set even
if a dipole is not requested, but the ccdensity module is run.