Adds setter for OPDM to wfn adds options for denscubeprops.
There are several post-scf methods with no wfn, or a severely limited
one. This change allows those methods to change the OPDM of the
reference wfn so that any input file side workup ie
DensityCubeProperties/CubeProperties will have the post scf density
matrix to use.

This hopefully will be temporary, and eventually wavefunction for all
methods will be implemented with the ability to properly modify
themselves.